+++ mayan/apps/mimetype/literals.py
@@ -1,3 +1,0 @@
-from __future__ import unicode_literals
-
-DEFAULT_MIMETYPE_FILE_READ_SIZE = 1024