+++ HISTORY.rst
@@ -94,6 +94,7 @@
   management app.
 * Move the purge permission logic to the StorePermission
   manager.
+* Remove the MIMETYPE_FILE_READ_SIZE setting.
 
 3.1.11 (2019-04-XX)
 ===================